@@ -27,30 +27,28 @@
   "dependencies": {
     "@babel/core": "^7.21.4",
     "@babel/runtime": "^7.17.9",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-jss": "^10.0.4",
-    "react-relay": "^11.0.0",
-    "react-router-dom": "^5.1.2",
-    "relay-runtime": "^15.0.0"
+    "react-relay": "^16.2.0",
+    "react-router-dom": "^6.22.0",
+    "relay-runtime": "^16.2.0"
   },
   "devDependencies": {
-    "@types/react": "^16.9.19",
-    "@types/react-dom": "^16.9.5",
-    "@types/react-relay": "^7.0.3",
-    "@types/react-router-dom": "^5.1.3",
-    "@types/relay-runtime": "^8.0.6",
+    "@types/react": "^18.2.0",
+    "@types/react-dom": "^18.2.0",
+    "@types/react-relay": "^14.1.3",
+    "@types/relay-runtime": "^14.1.23",
     "babel-loader": "^8.0.6",
     "babel-plugin-relay": "^8.0.0",
-    "graphql": "^15.0.0",
+    "graphql": "^16.8.1",
     "graphql-compiler": "^1.7.0",
-    "relay-compiler": "^15.0.0",
-    "relay-config": "^8.0.0",
+    "relay-compiler": "^16.2.0",
     "source-map-loader": "^0.2.4",
-    "ts-loader": "^6.2.1",
-    "typescript": "^3.7.5",
-    "webpack": "^4.41.6",
-    "webpack-cli": "^3.3.11",
+    "ts-loader": "^9.5.0",
+    "typescript": "^5.3.3",
+    "webpack": "^5.90.1",
+    "webpack-cli": "^5.1.4",
     "webpack-merge": "^4.2.2"
   }
 }
